rename baomap to store
just so it is consistent with iroh-sync
