+++ iroh-bytes/Cargo.toml
@@ -37,7 +37,7 @@ serde-error = "0.1.2"
 smallvec = { version = "1.10.0", features = ["serde", "const_new"] }
 subtle = "2.4"
 thiserror = "1"
-tokio = { version = "1", features = ["fs"] }
+tokio = { version = "1" }
 tokio-util = { version = "0.7", features = ["io-util", "io", "rt"] }
 tracing = "0.1"
 tracing-futures = "0.2.5"
@@ -51,4 +51,4 @@ tokio = { version = "1", features = ["macros", "test-util"] }
 
 [features]
 default = ["flat-db"]
-flat-db = ["reflink-copy"]
+flat-db = ["reflink-copy", "tokio/fs"]
