+++ iroh-bytes/src/lib.rs
@@ -3,7 +3,6 @@
 #![deny(missing_docs, rustdoc::broken_intra_doc_links)]
 #![recursion_limit = "256"]
 
-pub mod baomap;
 pub mod get;
 pub mod hashseq;
 pub mod protocol;
