+++ iroh/src/client.rs
@@ -13,8 +13,8 @@ use anyhow::{anyhow, Result};
 use bytes::Bytes;
 use futures::stream::BoxStream;
 use futures::{SinkExt, Stream, StreamExt, TryStreamExt};
-use iroh_bytes::baomap::ValidateProgress;
 use iroh_bytes::provider::AddProgress;
+use iroh_bytes::store::ValidateProgress;
 use iroh_bytes::Hash;
 use iroh_bytes::{BlobFormat, Tag};
 use iroh_net::{key::PublicKey, magic_endpoint::ConnectionInfo, PeerAddr};
