+++ iroh/src/sync_engine/rpc.rs
@@ -3,7 +3,7 @@
 use anyhow::anyhow;
 use futures::{FutureExt, Stream};
 use iroh_bytes::{
-    baomap::Store as BaoStore,
+    store::Store as BaoStore,
     util::{BlobFormat, RpcError},
 };
 use iroh_sync::{store::Store, sync::Namespace};
